move PathInfoParser class into its own module
